1378: Addressed review comments
