+++ packages/indicators/src/Builder/CountEventBuilder/config/validate.ts
@@ -4,22 +4,7 @@
  */
 
 import { constructIsEmptyOrSync, hasContent, isAString, isPlainObject } from '@tupaia/utils';
-import { getExpressionParserInstance } from '../../../getExpressionParserInstance';
-import { assertDefaultValuesHaveAllowedTypesOrUndefined } from '../../validators';
-import { CountEventConfig } from './types';
-
-const assertAllDefaultsAreCodesInFormula = (
-  defaultValues: Record<string, unknown>,
-  config: CountEventConfig,
-) => {
-  const parser = getExpressionParserInstance();
-  const variables = parser.getVariables(config.formula);
-  Object.keys(defaultValues).forEach(code => {
-    if (!variables.includes(code)) {
-      throw new Error(`'${code}' is in defaultValues but not referenced in the formula`);
-    }
-  });
-};
+import { assertAllDefaultsAreCodesInFormula, assertDefaultValuesHaveAllowedTypesOrUndefined } from '../../validators';
 
 const assertDefaultValuesHaveAppropriateTypes = (defaultValues: Record<string, unknown>) => {
   assertDefaultValuesHaveAllowedTypesOrUndefined(defaultValues, ['number', 'string']);
