+++ packages/indicators/src/Builder/validators.ts
@@ -3,6 +3,8 @@
  * Copyright (c) 2017 - 2021 Beyond Essential Systems Pty Ltd
  */
 
+import { getExpressionParserInstance } from '../getExpressionParserInstance';
+
 export const assertDefaultValuesHaveAllowedTypesOrUndefined = (
   defaultValues: Record<string, unknown>,
   allowedTypes: string[],
@@ -12,6 +14,19 @@ export const assertDefaultValuesHaveAllowedTypesOrUndefined = (
       throw new Error(
         `Value '${code}' in defaultValues is not in types ${allowedTypes.toString()} or 'undefined': ${value}`,
       );
+    }
+  });
+};
+
+export const assertAllDefaultsAreCodesInFormula = (
+  defaultValues: Record<string, unknown>,
+  config: Record<string, unknown>,
+) => {
+  const parser = getExpressionParserInstance();
+  const variables = parser.getVariables(config.formula);
+  Object.keys(defaultValues).forEach(code => {
+    if (!variables.includes(code)) {
+      throw new Error(`'${code}' is in defaultValues but not referenced in the formula`);
     }
   });
 };
